Working batteries and debug commands
Added in a chargeable item system.

Chargeable items can swap models when power reaches a certain percetage

Added in add, remove and set power commands that mark the world
as cheated when you enable debug mode

Made sure the client is brought up to speed when a value change occurs
server side. This happens once a second and only if a value has changed
to save on bandwith