Water mill and switches
Switches now have the ability to start and stop the flow of power. But
this affects a 2F radius from the switch so will need to look into how
to make that a bit more friendly and not be affected by the switch.
Maybe override the switch and make it it's own item that tells the
connecting wire that it has a switch to listen to. Will have to decide
sometime later

Water mills can now produce power. The water wheel defining the
efficiency of the water mill. The one included in the project produces
30% of 30 per cycle